+++ src/Inceptum.Messaging/ProcessingGroup.cs
@@ -1,8 +1,5 @@
 using System;
-using System.Collections.Concurrent;
 using System.Collections.Generic;
-using System.Linq;
-using System.Reactive.Concurrency;
 using System.Reactive.Disposables;
 using System.Runtime.Serialization;
 using System.Threading;
